requirements.txt
absl-py==1.4.0
astunparse==1.6.3
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.26.4
beautifulsoup4==4.12.2
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.5.4
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
imutils==0.5.4
install==1.3.5
isodate==0.6.1
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
msrest==0.7.1
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
opencv-contrib-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
pdf2image==1.16.3
Pillow==9.5.0
platformdirs==3.5.1
progressbar2==4.2.0
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.5.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.4.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uuid==1.30
Werkzeug==2.3.3
wrapt==1.15.0